LayoutTests/http/tests/media/media-source/video-media-source-append-with-offset-in-ended-state-expected.txt

Tests setting a timestampOffset and calling append() after the MediaSource has transitioned to the "ended" state.

Appending initialization segment.
EVENT(loadedmetadata)
Appending first media segment.
Calling endOfStream().
Setting timestampOffset to 1.
Appending second media segment.
END OF TEST
